app-icon

+++ db/migrate/20190624105330_add_andriod_version_to_versions.rb
@@ -1,5 +1,5 @@
 class AddAndriodVersionToVersions < ActiveRecord::Migration[5.2]
   def change
-    add_column :versions, :AndriodVersion, :integer
+    add_column :versions, :AndriodVersion, :float
   end
 end
